Address PR review: rename segment, relocate hash, drop FilePartitioning source flag
- Rename `child_segment`/`segment` -> `current_task_id_suffix` (this task's
  own segment of the lineage path). Clearer per review.
- Move `get_deterministic_hash` to `nemo_curator/utils/hash_utils.py` so
  non-text modalities (FileGroupTask, dedup) don't pull in text deps. The
  text `writer/utils.py` re-exports it for backward compat.
- Drop `is_source_stage = True` from FilePartitioningStage; rely on
  Pipeline.build()'s "first stage is source" default, so users who put a
  different stage first (e.g. URL generation) get the right behavior.
- Fix dead-branch assertion in test_tasks.py (parent lineage is always ""
  -> filtered, so children are str(i), never "_i").
