+++ nemo_curator/utils/hash_utils.py
@@ -1,0 +1,25 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import hashlib
+
+
+def get_deterministic_hash(inputs: list[str], seed: str = "") -> str:
+    """Create a deterministic hash from inputs.
+
+    Lives in ``nemo_curator.utils`` (not under ``stages/text``) so that
+    non-text modalities can use it without pulling in text dependencies.
+    """
+    combined = "|".join(sorted(inputs)) + "|" + seed
+    return hashlib.sha256(combined.encode()).hexdigest()[:12]
